requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
azure-functions
azure.storage.blob
pycryptodome==3.12.0
pandas==1.3.5
adal
pyodbc
azure-storage-file-share
mojimoji
